requirements.txt
DateTime
docxtpl
Flask
mysql-connector-python
openpyxl
flask-cors
python-docx
Pillow
gunicorn